Remove unused imports, remove unnecessary boxing of int-value and apply some other IDE suggestions
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1899076 13f79535-47bb-0310-9956-ffa450edef68

+++ poi/src/main/java/org/apache/poi/poifs/filesystem/POIFSFileSystem.java
@@ -454,7 +454,7 @@ public class POIFSFileSystem extends BlockStore
         // Ensure there's a spot in the file for it
         ByteBuffer buffer = ByteBuffer.allocate(bigBlockSize.getBigBlockSize());
         // Header isn't in BATs
-        long writeTo = Math.multiplyExact(1L + offset, (long)bigBlockSize.getBigBlockSize());
+        long writeTo = Math.multiplyExact(1L + offset, bigBlockSize.getBigBlockSize());
         _data.write(buffer, writeTo);
         // All done
         return newBAT;
